consolidating import APIs, refs#76
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@7432 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/netimport/NIXMLEdgesHandler.cpp
@@ -199,7 +199,7 @@ NIXMLEdgesHandler::myStartElement(SumoXMLTag element,
             return;
         }
         // check whether this lane exists
-        if (lane>=myCurrentEdge->getNoLanes()) {
+        if (lane>=(int) myCurrentEdge->getNoLanes()) {
             MsgHandler::getErrorInstance()->inform("Lane-id is larger than number of lanes (edge '" + myCurrentID + "').");
             return;
         }
